refactor(dashboard): extract Mastio core + settings + api status sub-routers (closes F-B-201, PR-13/13)
Move the final residual routes out of the mcp_proxy/dashboard/router.py
god-object into three dedicated modules:

- core_routes.py: smart entry / / org-status / display-name x3 / overview
- settings_routes.py: /settings page + password mgmt + license import
- api_status_routes.py: /api/update-status x2 + /api/version-status

The parent router.py is now a thin aggregator: prefix declaration plus
13 include_router calls (PR-2..PR-13). Helpers + template env factory
sit in _helpers.py and _template_env.py since PR-1.

Closes F-B-201 (BLOCKER, audit 2026-05-20). Mastio dashboard god-object
fully decomposed.
